fix: required cycle warning 해결
아래 warning 해결

```
Require cycle: src/components/home/oranges/OrangeBox.js -> src/components/home/oranges/Orange.js -> src/components/home/oranges/OrangeBox.js
```
